refactor(alerts-service): move makeAlertsManager to own file
